cleanup: move constant classes to appropriate files
